Warning compile flags and fixing warning (#1077)
* added warning flag to targets and error on warning; fixed then also all warning in code base

* addressing warnings and errors from CI

* addressing more warnings and errors

* MSVC really does not like unsigned int in omp parallel for

* some more error and warning fixes

* fixing some more warnings for MSVC build, not sure how to get rid of pybind11 warnings (should be disabled via pybind11.h header)

+++ examples/Cpp/ColorMapOptimization.cpp
@@ -47,7 +47,7 @@ int main(int argc, char *argv[]) {
                                       color_filenames);
     assert(depth_filenames.size() == color_filenames.size());
     std::vector<std::shared_ptr<geometry::RGBDImage>> rgbd_images;
-    for (int i = 0; i < depth_filenames.size(); i++) {
+    for (size_t i = 0; i < depth_filenames.size(); i++) {
         utility::LogDebug("reading {}...\n", depth_filenames[i]);
         auto depth = io::CreateImageFromFile(depth_filenames[i]);
         utility::LogDebug("reading {}...\n", color_filenames[i]);
