+++ src/Open3D/Geometry/PointCloudCluster.cpp
@@ -35,12 +35,13 @@
 namespace open3d {
 namespace geometry {
 
-std::vector<int> PointCloud::ClusterDBSCAN(double eps, int min_points) const {
+std::vector<int> PointCloud::ClusterDBSCAN(double eps,
+                                           size_t min_points) const {
     KDTreeFlann kdtree(*this);
     // set all labels to undefined (-2)
     std::vector<int> labels(points_.size(), -2);
     int cluster_label = 0;
-    for (int idx = 0; idx < points_.size(); ++idx) {
+    for (size_t idx = 0; idx < points_.size(); ++idx) {
         if (labels[idx] != -2) {  // label is not undefined
             continue;
         }
@@ -58,7 +59,7 @@ std::vector<int> PointCloud::ClusterDBSCAN(double eps, int min_points) const {
 
         std::unordered_set<int> nbs_next(nbs.begin(), nbs.end());
         std::unordered_set<int> nbs_visited;
-        nbs_visited.insert(idx);
+        nbs_visited.insert(int(idx));
 
         labels[idx] = cluster_label;
         while (!nbs_next.empty()) {
